virtual_lab_platform/requirements.txt
asgiref==3.8.1
Django==4.2.20
django-allauth==0.61.1
django-crispy-forms==2.1
crispy-bootstrap5==2023.10
python-dotenv==1.0.1
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2024.1
whitenoise==6.6.0
markdown==3.5.1
Pillow==10.0.0
